Make BatchedReceiver robust against race conditions
We found that we'd occasionally find race conditions. It seems that the server pointer (or it's 'subscriptions' field) we getting changed between the definition of the defer and its execution. By wrapping the call to AddInt64 in a function, the pointers are not evaluated until the defer is actually executed

Signed-off-by: Andrew Poydence <apoydence@pivotal.io>