trafficcontroller: use modified DNS resolver for log-cache client
There is a race condition between Bosh DNS having the log-cache DNS entry and when the traffic controller starts. The DNS resolver provided by gRPC only updates every 30 minutes, so in the case that the DNS entry is not available when the TC is restarted there will be ~30 minutes of down time.

This commit adds a copy of the DNS resolver from gRPC with changes to resolve the DNS every 10 seconds.

[#161989974]